refactor: remove code duplication
